Moved all code to main(). Moved files to Scripts folder.
